requirements.txt
reportlab~=4.0.5
Office365-REST-Python-Client==2.3.14
azure-functions
pandas
shapely